Creates a DataSet object and harmonise data against data

Creates a DataSet object and harmonise data against data

Usage

createDataSet(
  data = list(),
  ids,
  tools,
  source = "IEUopenGWAS",
  harmonise = TRUE,
  tolerance = 0.08,
  action = 1,
  strand = "forward",
  beta_col = "beta",
  se_col = "se",
  samplesize_col = "n",
  pvalue_col = "p",
  chr_col = "chr",
  position_col = "position",
  rsid_col = "rsid",
  effect_allele_col = "ea",
  other_allele_col = "nea",
  eaf_col = "eaf",
  id_col = "id",
  trait_col = "trait"
)

Arguments

data

A list of gwasglue2 SummarySet objects

ids

ID of GWAS studies to query

tools

Array of methods that gwasglue2 is going to convert the summarySet to (eg. "mr")

source

ID of Data Source to set the DataSet @slot metadata (Default and only option: IEUopenGWAS).

harmonise

logical (default TRUE). It harmonises the summary sets in the DataSet against each other.

tolerance

Inherited from harmoniseData() (default 0.08)

action

Inherited from harmoniseData() (Default 1)

strand

dna strand orientation (Default "forward", other option "reverse")

beta_col

Name of column with effect sizes. The default is "beta".

se_col

Name of column with standard errors. The default is "se".

samplesize_col

Column name for sample size. The default is "n".

pvalue_col

Name of column with p-value. The default is "p".

chr_col

Column name for chromosome . The default is "chr".

position_col

Column name for the position. Together, with @param chr gives the physical coordinates of the variant. The default is "position".

rsid_col

Required name of column with variants rs IDs. The default is "rsid".

effect_allele_col

Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G". The default is "ea".

other_allele_col

Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G". The default is "nea.

eaf_col

Name of column with effect allele frequency. The default is "eaf".

id_col

The default is "id".

trait_col

Column name for the column with phenotype name corresponding the the variant. The default is "trait"

Value

A harmonised gwasglue2 DataSet object